requirements.txt
langchain
langchain-core
langchain-community

faiss-cpu
sentence-transformers

pyyaml
pydantic
jsonschema

python-dotenv

groq
langchain-groq

networkx
pyvis

streamlit

sqlglot



